% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tbl_from_sql.R
\name{build_tbl_from_sql}
\alias{build_tbl_from_sql}
\title{Build a tbl object from defined SQL query}
\usage{
build_tbl_from_sql(sql, tablename = NULL)
}
\arguments{
\item{sql}{The SQL command to be executed, the resulted will be saved into a temporary table}

\item{tablename}{if not specified, a random name will be used}
}
\value{
returns dplyr tbl object
}
\description{
this function allows manually build tbl objects from input SQL commands
}

