% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{download remote tbl into local data.frame object}
\usage{
download(tbl, stringAsFactors = T)
}
\arguments{
\item{stringAsFactors}{convert columns of character-type into factors}

\item{x}{a tbl object}
}
\description{
download remote tbl into local data.frame object
}

