% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{head}
\alias{head}
\title{helper functions that make tbl objects like data.frames}
\usage{
head(x, ...)
}
\arguments{
\item{x}{see ?head for details, this can also be tbl objects}
}
\description{
an alternative head which works for tbl objects
}

