% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junyi_compute.R
\name{junyi.compute}
\alias{junyi.compute}
\title{A workable alternative for dplyr::compute to use in junyi bigquery datasets}
\usage{
junyi.compute(x, tablename = NULL, dataset.name = NULL)
}
\arguments{
\item{x}{a tbl object to be computed}

\item{tablename}{if not specified, then a random name will be generated}

\item{dataset.name}{the dataset name, if not given, then the dataset parameters must be defined by set_dataset()}
}
\value{
returns dplyr tbl object
}
\description{
the dplyr's compute() function will not work due to bigquery not supporting CREATE TABLE methods, use this alternative instead
}

