% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junyi_left_join.R
\name{junyi.left_join}
\alias{junyi.left_join}
\title{A workable alternative for dplyr::left_join to use in junyi bigquery datasets}
\usage{
junyi.left_join(x, y, by = NULL, by.x = NULL, by.y = NULL,
  restore.names = T, tablename = NULL)
}
\arguments{
\item{x}{tbl object}

\item{y}{tbl object}

\item{by}{join by column, can take a vector of names. eg. c("keyID","date)}

\item{by.x}{join by column in argument x}

\item{by.y}{join by column in argument y}

\item{tablename}{if not specified, then a random name will be generated}
}
\value{
returns dplyr tbl object
}
\description{
A workable alternative for dplyr::left_join to use in junyi bigquery datasets
}

