% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junyi_semi_join.R
\name{junyi.semi_join}
\alias{junyi.semi_join}
\title{A workable alternative for dplyr::semi_join to use in junyi bigquery datasets}
\usage{
junyi.semi_join(x, y, by = NULL, by.x = NULL, by.y = NULL,
  tablename = NULL)
}
\arguments{
\item{x}{tbl object}

\item{y}{tbl object}

\item{by}{join by column, can only take one value}

\item{by.x}{join by column in argument x}

\item{by.y}{join by column in argument y}

\item{tablename}{if not specified, then a random name will be generated}
}
\value{
returns dplyr tbl object
}
\description{
return all rows from x where there are matching values in y, keeping just columns from x.
}

