% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junyi_union.R
\name{junyi.union}
\alias{junyi.union}
\title{A workable alternative for dplyr::union to use in junyi bigquery datasets}
\usage{
junyi.union(...)
}
\arguments{
\item{...}{tbl objects to perform set union}
}
\value{
returns dplyr tbl object
}
\description{
A workable alternative for dplyr::union to use in junyi bigquery datasets
}

