% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_primary_tbls.R
\name{load_primary_tbls}
\alias{load_primary_tbls}
\title{load all pre-built tables}
\usage{
load_primary_tbls(backup.dataset.date = NULL, source.dataset = "tidy_data",
  tmp.dataset = "tidy_tmp")
}
\arguments{
\item{backup.dataset.date}{the latest date of the dataset junyi_yymmdd on bigquery. If not specified, then the most recent previous Monday's date will be used.}

\item{source.dataset}{the dataset where the pre-built tidy tbls are located}

\item{tmp.dataset}{a temporary dataset in which user has access to read and write tables}
}
\value{
returns dplyr tbl object
}
\description{
load all pre-built tables
}

