% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tbl_by_name.R
\name{load_tbl_by_name}
\alias{load_tbl_by_name}
\title{load tbl by tablename}
\usage{
load_tbl_by_name(tablename, dataset.name = NULL)
}
\arguments{
\item{tablename}{the table name to be loaded}

\item{dataset.name}{the dataset name, if not given, then the dataset parameters must be defined by set_dataset()}
}
\value{
returns dplyr tbl object
}
\description{
load tbl by tablename
}

