% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_sql.R
\name{mutate_sql}
\alias{mutate_sql}
\title{compute SQL command and insert a column into a tbl object}
\usage{
mutate_sql(x, sql)
}
\arguments{
\item{x}{tbl object}

\item{sql}{sql command to be executed}
}
\description{
can manually insert sql command in which is not supported by dplyr, such as bigquery's window functions
}
\examples{
\dontrun{
mutate_sql("LAG(date,1) OVER (PARTITION BY user_primary_key ORDER BY date) AS lag_date")
}
}
\references{
\url{https://cloud.google.com/bigquery/docs/reference/legacy-sql#windowfunctions}

\url{https://cran.rstudio.com/web/packages/dplyr/vignettes/window-functions.html}
}

