% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\name{nrow}
\alias{nrow}
\title{helper functions that make tbl objects like data.frames}
\usage{
nrow(x)
}
\arguments{
\item{x}{can be a tbl object, data frame or SQL statement}
}
\description{
an alternative nrow which works for tbl objects
}

