% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dataset.R
\name{set_dataset}
\alias{set_dataset}
\title{set a global destination dataset for temporary tables}
\usage{
set_dataset(backup.dataset.date, destination.dataset.name)
}
\arguments{
\item{backup.dataset.date}{weekly backup dataset date, the latest date of the dataset junyi_yymmdd on bigquery, eg. eg. "2017-1-9"}

\item{destination.dataset.name}{a temporary dataset in which user has access to read and write tables}
}
\description{
set a global destination dataset for temporary tables
}

