% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{helper functions that make tbl objects like data.frames}
\usage{
summary(x, column)
}
\arguments{
\item{x}{see ?summary for details, this can also be tbl objects}

\item{column}{the column containing values to be calculated}
}
\description{
an alternative summary which works for tbl objects
}

