% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{helper functions that make tbl objects like data.frames}
\usage{
table(x, column, n = 5000)
}
\arguments{
\item{x}{see ?table for details, this can also be tbl objects}

\item{column}{the column containing values to be calculated}

\item{n}{how many items to show, by default is 5000.}
}
\description{
an alternative base::table which works for tbl objects
}

