% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_info_Content.R
\name{tbl_info_Content}
\alias{tbl_info_Content}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_info_Content(force = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}
}
\value{
returns dplyr tbl object
}
\description{
junyi content's specs data, including categories and type
}

