% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_info_User.R
\name{tbl_info_User}
\alias{tbl_info_User}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_info_User(force = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}
}
\value{
returns dplyr tbl object
}
\description{
user's background data, the primary key of this table is user_primary_key
}
\details{
while most variables are unique for each user, fields userRole, userCity, userSchool exist multiple values, in which are separated by comma ","
}

