% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_log_AssignMission.R
\name{tbl_log_AssignMission}
\alias{tbl_log_AssignMission}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_log_AssignMission(force = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}
}
\value{
returns dplyr tbl object
}
\description{
log data of mission assignment, the minimal time unit for each row is every 10 mins. The primary key for each mission is missionID
}

