% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_log_DoingMission.R
\name{tbl_log_DoingMission}
\alias{tbl_log_DoingMission}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_log_DoingMission(force = F, load.only = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}

\item{load.only}{set to true to directly load pre-built tbl without building}
}
\value{
returns dplyr tbl object
}
\description{
log data of doing mission contents, the minimal time unit for each row is every 10 mins. The primary key for each mission is missionID, for each content is contentID
}

