% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_log_ProbQuiz.R
\name{tbl_log_ProbQuiz}
\alias{tbl_log_ProbQuiz}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_log_ProbQuiz(force = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}
}
\value{
returns dplyr tbl object
}
\description{
detailed log data of problem(exercise) usage, the unit of each row is each prob's quiz
}

