% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_log_VideoPlay.R
\name{tbl_log_VideoPlay}
\alias{tbl_log_VideoPlay}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_log_VideoPlay(force = F, load.only = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}

\item{load.only}{set to true to directly load pre-built tbl without building}
}
\value{
returns dplyr tbl object
}
\description{
log data of video play/reverse/forwarding
}

