% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_log_VideoProb.R
\name{tbl_log_VideoProb}
\alias{tbl_log_VideoProb}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_log_VideoProb(force = F, window.begin = NULL, window.end = NULL)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}

\item{window.begin}{the date to begin watching, in format yy/mm/dd}

\item{window.end}{the date to end watching, in format yy/mm/dd}
}
\value{
returns dplyr tbl object
}
\description{
log data of video and problem(exercise) usage, the minimal time unit for each row is every 10 mins.
}

