% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_relation_prob_quiz.R
\name{tbl_relation_prob_quiz}
\alias{tbl_relation_prob_quiz}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_relation_prob_quiz(force = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}
}
\value{
returns dplyr tbl object
}
\description{
relationships between problem(exercise) and it's sub quizes
}

