% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_relation_users_bySameIP.R
\name{tbl_relation_users_bySameIP}
\alias{tbl_relation_users_bySameIP}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_relation_users_bySameIP(force = F, window.begin = NULL,
  window.end = NULL)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}

\item{window.begin}{the date to begin watching, in format yy/mm/dd}

\item{window.end}{the date to end watching, in format yy/mm/dd}
}
\value{
returns dplyr tbl object
}
\description{
users which are from the same IPs used by others during the same day at the same hour (aka. simultaneous users)
}

