% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_relation_users_SS_byClass.R
\name{tbl_relation_users_SS_byClass}
\alias{tbl_relation_users_SS_byClass}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_relation_users_SS_byClass(force = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}
}
\value{
returns dplyr tbl object
}
\description{
student and student relationships, defined by classID
}

