% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_relation_users_SS_byCoach.R
\name{tbl_relation_users_SS_byCoach}
\alias{tbl_relation_users_SS_byCoach}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_relation_users_SS_byCoach(force = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}
}
\value{
returns dplyr tbl object
}
\description{
student and student relationships, defined by coach-trainee relationship (having a common teacher)
}

