% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_relation_content_proximity.R
\name{tbl_relationship_content_proximity}
\alias{tbl_relationship_content_proximity}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_relationship_content_proximity(force = F, load.only = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}

\item{load.only}{set to true to directly load pre-built tbl without building}
}
\value{
returns dplyr tbl object
}
\description{
the shelf proximity of different contents that display at the Junyi's website
}

