% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{unique}
\alias{unique}
\title{helper functions that make tbl objects like data.frames}
\usage{
unique(x)
}
\arguments{
\item{x}{tbl object}
}
\description{
an alternative unique which works for tbl objects
}

