% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{helper functions that make tbl objects like data.frames}
\usage{
view(x, title, n = 5000, func = c("view", "head", "colnames"))
}
\arguments{
\item{x}{see ?View for details, this can also be tbl objects}

\item{title}{see ?View for details}

\item{n}{how many rows to show, by default is 1000. Can be set to Inf if needed}

\item{func}{shows view or prints out head or returns colnames}
}
\description{
an alternative View which works for tbl objects
}

