% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_primary_tbls.R
\name{build_primary_tbls}
\alias{build_primary_tbls}
\title{build all tidy tbls}
\usage{
build_primary_tbls(force = F, load.only = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}

\item{load.only}{set to true to directly load pre-built tbls without building them}
}
\value{
returns dplyr tbl object
}
\description{
build all tidy tbls
}

