% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junyi_compute.R
\name{junyi.compute}
\alias{junyi.compute}
\title{A workable alternative for dplyr::compute to use in junyi bigquery datasets}
\usage{
junyi.compute(x, tablename = NULL)
}
\arguments{
\item{x}{a tbl object to be computed, the results will be saved into a temporary dataset}

\item{tablename}{if not specified, then a random name will be generated}
}
\value{
returns dplyr tbl object
}
\description{
the dplyr's compute() function will not work due to bigquery not supporting CREATE TABLE methods, use this alternative instead
}

