% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_relation_mission_content.R
\name{tbl_relation_mission_content}
\alias{tbl_relation_mission_content}
\title{create tbl objects of various junyi bigquery datasets}
\usage{
tbl_relation_mission_content(force = F, load.only = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}

\item{load.only}{set to true to directly load pre-built tbl without building}
}
\value{
returns dplyr tbl object
}
\description{
relationships between contents and missions. The primary key for each mission is missionID, for each content is contentID
}

