% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_relation_user_identifiers.R
\name{tbl_relation_user_identifiers}
\alias{tbl_relation_user_identifiers}
\title{creates a tbl containing various user identifiers}
\usage{
tbl_relation_user_identifiers(force = F, load.only = F)
}
\arguments{
\item{force}{create a new data table despite a previous table with a same name already exist}

\item{load.only}{set to true to directly load pre-built tbl without building}
}
\value{
returns dplyr tbl object
}
\description{
the primary key of this table is user_primary_key, however, there exist multiple variables describing such user
}

